implementation of header files as static library and reorganization of directories
