+++ src/ElectricLib.cpp
@@ -1,4 +1,4 @@
-#include "ElectricLib.h"
+#include "includes/ElectricLib.h"
 
 #ifndef MAT_PI
 #    define MAT_PI 3.14159265358979323846
