+++ src/Tethys1DLib.cpp
@@ -1,4 +1,4 @@
-#include "Tethys1DLib.h"
+#include "includes/Tethys1DLib.h"
 
 
 using namespace H5;
