+++ src/TethysLib.cpp
@@ -1,4 +1,4 @@
-#include "TethysLib.h"
+#include "includes/TethysLib.h"
 
 
 using namespace H5;
